# students/K3341/Zotov_Mikhail/reports/docs/Lr1/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Hello","operationId":"hello__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/genres/":{"get":{"tags":["Genres"],"summary":"Read Genres","operationId":"read_genres_genres__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/GenreRead"},"type":"array","title":"Response Read Genres Genres  Get"}}}}}},"post":{"tags":["Genres"],"summary":"Create Genre","operationId":"create_genre_genres__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/genres/{genre_id}":{"get":{"tags":["Genres"],"summary":"Read Genre","operationId":"read_genre_genres__genre_id__get","parameters":[{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Genres"],"summary":"Update Genre","operationId":"update_genre_genres__genre_id__patch","parameters":[{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Genres"],"summary":"Delete Genre","operationId":"delete_genre_genres__genre_id__delete","parameters":[{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/":{"get":{"tags":["Users"],"summary":"Read Users","operationId":"read_users_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserRead"},"type":"array","title":"Response Read Users Users  Get"}}}}}}},"/users/me":{"get":{"tags":["Users"],"summary":"Read Current User","operationId":"read_current_user_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}}},"security":[{"HTTPBearer":[]}]},"patch":{"tags":["Users"],"summary":"Update Current User","operationId":"update_current_user_users_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/users/me/books":{"get":{"tags":["Users"],"summary":"Read Current User Books","operationId":"read_current_user_books_users_me_books_get","security":[{"HTTPBearer":[]}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookRead"},"title":"Response Read Current User Books Users Me Books Get"}}}}}},"post":{"tags":["Users"],"summary":"Add Book To Current User","operationId":"add_book_to_current_user_users_me_books_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"book_id","in":"query","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/books":{"get":{"tags":["Users"],"summary":"Read User Books","operationId":"read_user_books_users__user_id__books_get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookRead"},"title":"Response Read User Books Users  User Id  Books Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me/books/{book_id}":{"patch":{"tags":["Users"],"summary":"Update Book Availability","operationId":"update_book_availability_users_me_books__book_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LibraryUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBookRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users"],"summary":"Remove Book From Current User","operationId":"remove_book_from_current_user_users_me_books__book_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"tags":["Users"],"summary":"Read User","operationId":"read_user_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Users"],"summary":"Update User","operationId":"update_user_users__user_id__patch","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users"],"summary":"Delete User","operationId":"delete_user_users__user_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["Auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/":{"get":{"tags":["Books"],"summary":"Read Books","operationId":"read_books_books__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BookRead"},"type":"array","title":"Response Read Books Books  Get"}}}}}},"post":{"tags":["Books"],"summary":"Create Book","operationId":"create_book_books__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/{book_id}":{"get":{"tags":["Books"],"summary":"Read Book","operationId":"read_book_books__book_id__get","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Books"],"summary":"Update Book","operationId":"update_book_books__book_id__put","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Books"],"summary":"Delete Book","operationId":"delete_book_books__book_id__delete","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/exchanges/":{"post":{"tags":["Exchange Requests"],"summary":"Create Exchange Request","operationId":"create_exchange_request_exchanges__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExchangeRequestCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExchangeRequestRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/exchanges/me/sent":{"get":{"tags":["Exchange Requests"],"summary":"Get Sent Requests","operationId":"get_sent_requests_exchanges_me_sent_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ExchangeRequestRead"},"type":"array","title":"Response Get Sent Requests Exchanges Me Sent Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/exchanges/me/received":{"get":{"tags":["Exchange Requests"],"summary":"Get Received Requests","operationId":"get_received_requests_exchanges_me_received_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ExchangeRequestRead"},"type":"array","title":"Response Get Received Requests Exchanges Me Received Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/exchanges/{request_id}/respond":{"post":{"tags":["Exchange Requests"],"summary":"Respond To Exchange Request","operationId":"respond_to_exchange_request_exchanges__request_id__respond_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"integer","title":"Request Id"}},{"name":"decision","in":"query","required":true,"schema":{"$ref":"#/components/schemas/StatusExchangeRequest"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExchangeRequestRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/exchanges/{request_id}":{"delete":{"tags":["Exchange Requests"],"summary":"Delete Exchange Request","operationId":"delete_exchange_request_exchanges__request_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"integer","title":"Request Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BookCreate":{"properties":{"title":{"type":"string","title":"Title"},"author":{"type":"string","title":"Author"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"isbn":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Isbn"}},"type":"object","required":["title","author"],"title":"BookCreate"},"BookRead":{"properties":{"title":{"type":"string","title":"Title"},"author":{"type":"string","title":"Author"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"isbn":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Isbn"},"id":{"type":"integer","title":"Id"},"genres":{"items":{"$ref":"#/components/schemas/Genre"},"type":"array","title":"Genres"}},"type":"object","required":["title","author","id","genres"],"title":"BookRead"},"BookUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"author":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"isbn":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Isbn"}},"type":"object","title":"BookUpdate"},"ExchangeRequestCreate":{"properties":{"requested_book_id":{"type":"integer","title":"Requested Book Id"},"offered_book_id":{"type":"integer","title":"Offered Book Id"},"owner_id":{"type":"integer","title":"Owner Id"}},"type":"object","required":["requested_book_id","offered_book_id","owner_id"],"title":"ExchangeRequestCreate"},"ExchangeRequestRead":{"properties":{"id":{"type":"integer","title":"Id"},"status":{"$ref":"#/components/schemas/StatusExchangeRequest"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"requester":{"$ref":"#/components/schemas/UserRead"},"owner":{"$ref":"#/components/schemas/UserRead"},"requested_book":{"$ref":"#/components/schemas/BookRead"},"offered_book":{"$ref":"#/components/schemas/BookRead"}},"type":"object","required":["id","status","created_at","requester","owner","requested_book","offered_book"],"title":"ExchangeRequestRead"},"Genre":{"properties":{"name":{"type":"string","title":"Name"},"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"}},"type":"object","required":["name"],"title":"Genre"},"GenreCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"GenreCreate"},"GenreRead":{"properties":{"name":{"type":"string","title":"Name"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","id"],"title":"GenreRead"},"GenreUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},"type":"object","title":"GenreUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LibraryUpdate":{"properties":{"is_available":{"type":"boolean","title":"Is Available"}},"type":"object","required":["is_available"],"title":"LibraryUpdate"},"StatusExchangeRequest":{"type":"string","enum":["pending","accepted","declined"],"title":"StatusExchangeRequest"},"UserBookRead":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"author":{"type":"string","title":"Author"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"isbn":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Isbn"},"is_available":{"type":"boolean","title":"Is Available"}},"type":"object","required":["id","title","author","is_available"],"title":"UserBookRead"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","email","password"],"title":"UserCreate"},"UserLogin":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserLogin"},"UserRead":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"bio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bio"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["username","email","id"],"title":"UserRead"},"UserUpdate":{"properties":{"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"},"bio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bio"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}